moved logout route to source_app/views.py
