+++ securedrop/source_templates/base.html
@@ -30,7 +30,7 @@
 
       <div class="panel selected">
       {% if 'logged_in' in session %}
-        <a href="{{ url_for('logout') }}" class="sd-button btn pull-right" id="logout">{{ gettext('LOG OUT') }}</a>
+        <a href="{{ url_for('main.logout') }}" class="sd-button btn pull-right" id="logout">{{ gettext('LOG OUT') }}</a>
       {% endif %}
         <hr class="no-line">
